refactor(info pop up): removed duplicated code by creating an abstract class
