+++ test/views/pages/profile/profile_page_test.dart
@@ -6,6 +6,7 @@ import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/models/database/post/post_firestore.dart";
 import "package:proxima/models/database/user/user_firestore.dart";
 import "package:proxima/services/database/comment/comment_repository_service.dart";
+import "package:proxima/views/components/content/info_pop_up.dart";
 import "package:proxima/views/components/content/user_avatar/user_avatar.dart";
 import "package:proxima/views/navigation/leading_back_button/leading_back_button.dart";
 import "package:proxima/views/pages/home/home_top_bar/home_top_bar.dart";
@@ -128,12 +129,11 @@ void main() {
       expect(postPopup, findsOneWidget);
 
       //Check that the title of the pop up is displayed
-      final postPopupTitle = find.byKey(ProfileInfoPopUp.popUpTitleKey);
+      final postPopupTitle = find.byKey(InfoPopUp.popUpTitleKey);
       expect(postPopupTitle, findsOneWidget);
 
       //Check that the description of the pop up is displayed
-      final postPopupDescription =
-          find.byKey(ProfileInfoPopUp.popUpDescriptionKey);
+      final postPopupDescription = find.byKey(InfoPopUp.popUpDescriptionKey);
       expect(postPopupDescription, findsOneWidget);
 
       // Check that post content is displayed on popup
@@ -149,8 +149,7 @@ void main() {
       expect(descriptionContent, findsOneWidget);
 
       //Check that the delete button is displayed
-      final postPopupDeleteButton =
-          find.byKey(ProfileInfoPopUp.popUpDeleteButtonKey);
+      final postPopupDeleteButton = find.byKey(ProfileInfoPopUp.popUpButtonKey);
       expect(postPopupDeleteButton, findsOneWidget);
 
       //Check clicking on the delete button come back to the profile page
@@ -184,13 +183,12 @@ void main() {
       expect(commentPopup, findsOneWidget);
 
       //Check that the description of the pop up is displayed
-      final commentPopupDescription =
-          find.byKey(ProfileInfoPopUp.popUpDescriptionKey);
+      final commentPopupDescription = find.byKey(InfoPopUp.popUpDescriptionKey);
       expect(commentPopupDescription, findsOneWidget);
 
       //Check that the delete button is displayed
       final commentPopupDeleteButton =
-          find.byKey(ProfileInfoPopUp.popUpDeleteButtonKey);
+          find.byKey(ProfileInfoPopUp.popUpButtonKey);
       expect(commentPopupDeleteButton, findsOneWidget);
 
       //Check clicking on the delete button come back to the profile page
